+++ backpack.tf/content.js
@@ -2,7 +2,6 @@
 File for routing the scripts for backpack.tf based
 on the hostname and pathaname the user is in
 */
-//import { createKSButtons } from "./oldUI/addKSButton/content.js";
 
 /**
 @description: Function to route the pathname to the scripts
@@ -16,7 +15,7 @@ function scriptRouter() {
   } else {
     /* start scripts for oldUI */
     //createKSButtons(url);
-    loadAddKSButtons()
+    loadAddKSButtons(url)
   }
 }
 
@@ -26,14 +25,13 @@ scriptRouter();
 /* 
 Utility funtions to load scripts
 */
-function loadAddKSButtons() {
+function loadAddKSButtons(url) {
   (async () => {
     // Load module dynamically
     const { createKSButtons } = await import(
       chrome.runtime.getURL("backpack.tf/oldUI/addKSButtons/content.js")
     );
 
-    const url = new URL(location.href);
     createKSButtons(url);
   })();
 }