+++ steamcommunity.com/changeLog.md
@@ -1,3 +1,8 @@
+# Version 1.0.1
+
+Forked steamTradeHelper to steamTradeHelper script page
+Added routerof scripts in content.js
+
 # Version 1.0.0
 
 Added steamLinks